requirements.txt
# Core dependencies
questionary==2.0.1
typer==0.9.0
streamlit==1.28.1
pyyaml==6.0.1
pandas==2.1.3
python-dotenv==1.0.0

# CLI enhancements
rich==13.7.0
click==8.1.7

# Data handling
jsonlines==4.0.0

# API interactions
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: API documentation parsing
prance==23.6.21.0